translations, move to flashcards components
